Update todo.md
I've spent too much time finding problems in entries so I've added
specifically what entries have to be checked. If I hadn't done this I
would be done with format fixes in README.md (at least most likely). I
might end up still trying to find if they work but I should probably
wait on it so this should be added to the todo, to check entries in 2011
(some are already known to be a problem and are being worked on but the
last two, vic and zucker, have not been checked at all - but neither
have their README.md files). Of course I might check entries anyway as
it is useful to do it with the README.md files but I need to try and
focus on the task - format fixes of the README.md files. The locating
and fixing entries can come later.