Format fix README.md + add alt code to 2011/vik
The alternate code is based on the author's statements on how to get it
to work for Windows. I then looked at MS docs to find out what header
files are needed. The only reason I did this is because the purpose of
the issue #1 is to get the entries to compile under as many systems as
possible and since the author gave instructions ... Obviously I didn't
test this as I have no way to do so. It's 'as is'. In the famous words
of Douglas Adams: 'Don't share and don't enjoy! :-)'

Updated bugs.md with INABIAF noted by the author.

Format fix README.md.