Format fix 2012/vik/README.md + add alt code
Based on the author's description I made an alt version that should
theoretically work in Windows though of course (thank goodness) I have
no way to actually test it.

Just like with 2011/vik I've only done this because part of issue #1 is
to make entries work on as many systems as possible though it's not
entirely clear to me that 'Windows' should even be considered a 'system'
at all let alone one worthy of consideration :-)

Updated .gitignore.