Address 2011/richards in bugs.md wrt Apple M1
I've kind of wondered for a while if it's due to Apple's restrictions
against memory protection (in mmap() wrt exec/write/read) and it seems
like that this might be the case either for the Apple silicon chip or
others too (I can only test the M1 chip).

I linked to resources found on Apple's website as well as added a bit of
information I discovered but I think at this time this cannot be
resolved easily or at all because Apple seems to think they know better
than everyone else on certain things and unfortunately they make it very
hard to work round these types of decisions. I will proceed to other
years now, starting with 2013/birken, in the next commits ... unless I
find a typo or something in an earlier year prior to continuing 2013 :-)