Get 2014/skeggs to compile
The problem was that the CDEFINE was missing 's: the define is an
actual string so it needs the 's surrounding the "s otherwise the code
will not compile.

I didn't have a chance to really look at the compilation error but I
noticed in Yusuke Endoh's analysis that he compiled it slightly
different than the Makefile. It became clear at that point what was
going on. He added -ldl but this should not be necessary so I did not
add it to the Makefile.