Typo and format fix 1995/README.md
Funny and ironic fact: due to cold hands (unfortunately fan is on but I
can't turn it off due to a cat between my legs)  when I created a new branch
for this (which has typo fixes) I initially left off a 9 in 1995. So
there was almost a tpyo for a tpyo fix cmoimt.