Fix winners.sql, remove prizes.sql
The fixes to winners.sql will be important for the soon (it is to be
hoped as I know how to do it now) to be completed author-wins.csv file.
It is the next thing I plan to work on hopefully later today.

The prizes.sql in additional to also having mistakes is not needed so
it's been deleted.