Completely process author to author wins
This addresses the problem recently brought up in issue #4.

I should point out that this should be rechecked just to be sure
something didn't slip through as there were a few times that I joined a
line and then noticed it was a similar name (like first name but
different surname) and then undid the join. However I think we're good:
assuming that the SQL file was entirely correct which as we know there
were a few problems I had to fix.

One was strange - it had html in it. This one also had the wrong name:
should be Mary Ann Horton, not Mark Horton as she changed gender.
Previously it was something like Mark (Mary) Horton but I fixed that -
but not in the SQL file as it was out of the repo and I totally forgot
about it. The sql file winners.sql has been updated accordingly.