Fix fix on 2004/gavin/fs.tar
The tarball, whilst in a subdirectory, did not extract into a
subdirectory so that if one did tar xvf img/fs.tar it would still
overwrite the files. As well make clobber would wipe the files out. This
does mean one will have to update the tarball if they are to use it but
one presumes that they would know how to do this.

Note that kernel is still under img/ but it's not part of the tarball
and it's supposed to be outside a tarball too.