Make 2001/anonymous more like before wrt CPP
It now uses the G macro that was unused in the fix. I plan to look at
the other macros that were changed to be unused as well. The hope is I
can make it look much more like it originally looked but still let it
work. Last I remember unfortunately it's not simply a matter of changing
that function that uses the macros but it might be worth trying that
again.

Also in this commit I changed the literal "anonymous.c" to be __FILE__
like it was before.