Make 1985/sicherman more like original wrt CPP
Although the no longer valid code is commented out the macros and the
extra function (subr()) now exist to make it look more like it used to
before the fix.

I have another version to make it even more like the original but a
question that has to be answered is which should be the sicherman.c.
This can be decided at another time but I see it both ways: the way it
is now provides more educational value but it seems less 'authentic'
too. But on the other hand having it closer to the original with
commented out code could also be an issue - maybe. Even so the working
version might have some code that can be replaced with the macros to
make it more like before (this has yet to be determined).