Deprecate twitter
We further deprecate the use of twitter.

In cases such as historic guidelines and rules, we add
a note about using Mastodon.

In cases where a tweet URL was used, we replace with
an Internet Archive link because twitter has announced
that accounts that are not used for a period of time
will be deleted.

We add a few missing deprecated_twitter information
to the author.numbers and author.cvs files.

We remove references to twitter handles (after making
sure that both author.numbers and author.cvs has the
deprecated_twitter information).  When the "_README.md_"
files have their header information generated by
a tool, such deprecated_twitter data will not be present.