Slight update to 2019/poikola/README.md
Changed my name to my full name. I incidentally meant to say that yes I
do like the entry Timo but space, stars, the sky, planets,
constellations and all such things are much closer to my heart so the
Most stellar entry is my favourite of the two! But still this is a nice
one. Again thank you!