Fix Enigma chocolate cake files wrt twitter
With commit f46e68740fd35574bdcd8842c4d88b4df019fdcd twitter links were
converted to archived copies (due to Elon Musk threatening to delete
accounts that are not active) and the chocolate cake files had a link to
a tweet. However the 2020/ferguson2 chocolate cake files were encrypted
with a key (one of 16 I gave to experiment with the entry itself) which
means that the files were not updated. Now they have been so that when
properly decrypted they will link not to twitter itself but twitter on
the Internet Wayback Machine.

Question: what's the difference between SpaceX rockets and Twitter? Elon
Musk hasn't yet managed to blow up twitter!