Various updates in todo.md
What prompted this update is a 'problem' I noticed in 2019/duble: it
leaves sockets lying round in the current working directory but it might
not be safe for make clobber to delete them: involves '.' files and we
don't know if the glob will be different in different systems. In macOS
it appears to be '^\.[A-Z].*' (which in commit
cfb9c66fae54a7a83ce3de8aaf54df946bdabdef I made a mistake saying it
ended with '+*' which is obviously very wrong).

The README.md and bugs.md does show how to delete that pattern but it is
TBD if make clobber should do this.