Fix URL in 1994/ldb
Unfortunately the link is dead so we have to use the Internet Wayback
Machine until such a time that a new link is found. Perhaps this should
be added to the bugs.md with the status about dead links but this has
not been done for the websites which were archived. This can be
determined another time - or not.