Improve documentation of installing SDL and SDL2
Remove how to install SDL1 / SDL2 from var.mk as it's already in faq.md.
Instead refer to faq.md. How to compile (variables etc.) is still in the
var.mk file.

Reworded that section a bit and added 'Other linux distributions'
subsection to 'Linux' (in faq.md).

Minor typo fix in faq.md at the same time as the above change. Normally
I would not do this but it's one word changed and I am too tired at this
time (or can't be bothered?) to do another pull request for one word.