update archive to the 2020 Dec 28 set
As the GitHub repo was released on 2022 Dec 31,
the 2020 Dec 28 bzip2 compressed tarball set
represent a better archival point than 2022 Dec 29
fork of temp-test-ioccc repo.

Moved the `all/all.tar.bz2` to `achive/archive-all.tar.bz2`
as that file corresponds with what was in the
[2020 Dec 28](https://web.archive.org/web/20201228005211/http://www.ioccc.org/)
Internet Archive Wayback machine snapshot.

It is worth noting that the 2020 Dec 28 bzip2 compressed tarball
set had a different archive for both 1985 and 1992.

See the README.md file for a more detailed analysis of the archive.

We recommend that you update your uncompressed
archive set as follows:

```sh
for year in 198{4,5,6,7,8,9} 199{0,1,2,3,4,5,6,8} 200{0,1,4,5,6} 201{1,2,3,4,5,8,9} 2020; do
    rm -rf "$year"
    tar -jxvf "archive-$year.tar.bz2"
done
```