Add www-history.md, setup thanks-for-fixes.md
The `www-history.md` file has been added, in part, because we added
`archive/README.md` and wanted to capture some of the history of
how the IOCCC winners have been presented.  This file is **NOT**
a history of the IOCCC: it is more of a way to preserve some
of the history of the web site and repo.

The `thanks-for-fixes.md` started out as a `fix-thanks.md` file and
a `THANKS.md` file.  Both filenames were problematic.  This file
is **NOT** a way to thank the many people, past IOCCC winners as
just one example, who have made a positive difference to the IOCCC.
This file is **NOT** a place to document every time some winner
helps edit / fix another winner's entry, not is ti intended to be
a comprehensive list of every time someone pointed out a web site
problem or submitted a pull request to fix something.  The purpose
of this file is to acknowledge those who made improvements to the
way that IOCCC winners are presented.