Fix 2004/vik2 for Linux
Unfortunately this fix required heavy modification to the original file
and inclusion of a file that was once generated by the Makefile but now
it works for both macOS and Linux. It unfortunately will generate a lot
of warnings in Linux (or at least some - in Rocky Linux and Fedora Linux
it generates a lot of warnings about GCC extensions and with clang it
also generates warning(s?) about reserved keywords for a total of 45131
(!!) warnings) but still it now works with Linux and macOS (clean
compilation).

See the thanks-for-fixes.md file for example warnings and what this fix
entailed.