Remove last references to .hint files
This changes in years.html the 2011/zucker/hint.html to
2011/zucker/README.md (at the top of the file - to do with the IOCCC
logo) and also 1998/dorssel/dorssel.hint to 1998/dorssel/dorssel.md in
years.html (which should have been done in
ccbbacb31b7bb145b529380cada5af54ae33110e but was left out by accident).
As well it fixes 2004/gavin in bugs.md to refer to (in the Makefile rule
that is included) README.md.

There are still files with 'hint' in the name however. I believe these
to be correct.