Fix 2005/mikeash for linux
The problem was to do with an unknown escape sequence (\N) which was
used in a string (also in a comment and the 'N' was referred to in
code). It had been '\Newline' but this was changed to '\nNewline', the
array size that stores part of the string was increased by 1 and the
code was updated to check for 'n' instead of 'N'.

I also fixed a problem in README.md: might not have not might not of.