Change how 1985/lycklama is built + fix typo
Now that we have the original code in lycklama.orig.c we can have what
would more appropriately be an alternate version in lycklama.alt.c which
is now done.

Typo fix was really unfortunate: it was in an example command to try
which would have failed to work as the name was wrong.