Change bug status in 2006/toledo2 + download files
On another look I believe it to be a typo of the author that the prompt
should be 'A>'. I believe it should be what it is which is '>'. Still
there appears there might be a problem but because I am not sure of this
I changed it to possible bug. The problem or possible problem is that
the use of DDT.COM seems to not work but this might be not enough
information from the author or it might be a misunderstanding or
something else entirely.

I also downloaded the files that were referred to in the author's
remarks and added them to the repo. The ones from the SOURCE/ directory
are in the entry directory too as those might be required (though the
use might also be the possible or actual bug described above). This was
done to make it easier to use but also as part of one of the issues on
GitHub.