Fixes to 1984/decot
Make the decot.orig.c what was originally submitted prior to the judges
changing a line due to some compilers disliking that line. The alt
version does not have this fix applied. Note that I yanked paste the
line so if there was a typo that's there too though it appears not to
have that problem.

Updated Makefile noting that the purpose of the entry is to print out
rubbish so if one sees nonsense it's expected.

Fix README.md to refer to decot.orig.c instead of decot.alt.c for the
change made by the judges (that is to see it prior to the judges'
change).