Fixes and improvements to 2015/endoh3
Format fix and typo check 2015/endoh3.

Added rules to Makefile back_to and mullender to more easily enjoy the
theme of the entry. This undoes a previous change I made which made the
default rule compile mullender.c into prog which takes away another
thing of the entry when nothing is there.

Another pass of this entry will have to be done for typos/formatting but
I noticed the above problem with the Makefile and I wanted to fix it.

Fixed a typo in thanks-for-fixes.md as well.