Change bug status of 1992/kivinen
Although it maybe could be fixed and I might look at it in the future (I
would have to look at Xlib API first at the least) the fact that there
is a known workaround means that (based on discussion with Landon) it
can be considered a known issue that need not be fixed necessarily. Thus
it's now considered, at least tentatively, a feature not a bug. As for
what it is: when you start the game everything moves over to the right
of the screen and quickly ends but if you click on the game screen it
goes back to the centre to allow you to play.