More fixes in 2001/bellard
Fixed supplementary program to add a newline after printing the first
line: else the output of both lines were merged and hard to read.

Updated Makefile to always compile the supplementary program.

Modified thanks-for-fixes.md wrt this entry including fixing the link to
the supplementary program (it did not specify the directory: it just had
the file name itself).