Change bug status of 1994/ldb
The issue is that when fixing the interspersed output of the annoying
warning in macOS about using gets() an extra newline is printed out.
This is because fgets() stores the newline and gets() does not. But
since the warning kind of ruins the entry and a blank line is better
than the warning it seems like a worthy compromise. That was already
agreed upon but as it's a one liner and it's already quite long it might
be better to just not worry about it as although kind of annoying a
blank line isn't a big deal.