Improve my earlier fixes to 1990/tbr
The code now looks much more like the original. This was done through
the use of macros to redefine exit() and gets() so that the exit() call
uses the comma operator (,0) (because it is used in a binary expression)
and gets() now uses fgets() (details below).

The usual problem with fgets() versus gets() (and it was changed
primarily because with macOS the warning about gets() also occurs at
runtime and it is interspersed with the interactive output of this
program and this caused confusing output though a nice side effect is
it's safer too) is that fgets() retains the newline whereas gets() does
not (supposedly for backwards compatibility).

In some cases one can get away with not removing it (at the expense of a
newline possibly being printed) but that was not possible here. This
normally means one has to check for fgets() returning NULL which
required another binary expression (... != NULL || exit(0),0 ...) which
made the line much longer and not snakey. It appears that in this case
(possibly because it's used in a for loop but I did not examine this in
more detail) one can get away with not checking for NULL first and just
remove the newline so that the line looks much more like the original.