Improve top of thanks-for-fixes.md format a bit
I gave an example of one of the challenging bug fixes that Yusuke Endoh
made just like I have done for me.

I also made the fact that there are complicated fixes bold for both
Yusuke and me. Not strictly necessary but I am proud and I should be
too. Plus it's simply better this way.