Slight improvement to my fix of 1992/adrian
Originally there was one call to gets() and one call to fgets(). The
gets() was changed to the fgets() to prevent warnings including (in
macOS) warnings interspersed with output of the program.

But to make it more like the original gets() has been redefined to use
fgets(). However this does not mean that we can make the macro do a &&
... to strip off the newline as doing this will make a different file
fail to generate properly (I might look at that later as it would be
nicer if it looked exactly like before). Thus the 'Y[strlen(Y)-1]='\0';'
is still there in the body of the while loop at least for now.

The README.md was updated to remove the address - just have name and
country.

Updated thanks-for-fixes.md for the above change as well as some typo
fixes elsewhere.