Minor update to 1992/lush/README.md
Removed reference to original source file in the archive (this was
originally added as some changes had to be made to get it to work with
modern gcc - does not work with clang due to how it is programmed) as
the original is lush.orig.c. There are other README.md files for other
entries that have this same thing and these files will be updated in
time too.