Update bugs.md wrt 1995/cdua
At one point prior to fixing (for clang) 1995/cdua I noticed that
sometimes when running at runtime it would pause and ask you to press
return/enter again to resume. I have not noticed the last several times
(throughout many days) I have run it and that includes numerous runs
each day and so I have changed the status to possible bug, not
definitely a bug.

If I notice it again I will change it back or else look into the problem
but without it happening it's obviously harder to verify that it's a
problem (and since it's not every time it's also impossible to prove,
especially with obfuscated code, that it's fixed).