Fixes in 1998/schweikh1/README.md
There was 'I' in the README.md that was an artefact left over from
before the thanks file was introduced so that sentence was removed.

The thanks file was slightly updated too wrt this entry both as a new
category of fixes (the alternate code I made so that it works with macOS
but also referring to the README.md file).