Minor updates and fixes to ioccc_status.sh
Change -I IOCCC_status_version to -i IOCCC_status_version.

Change sed separator of sed operations to be '|', not '/'. There still
is the problem of if the version has a '|' in it the sed command will
break but this might be solved by requiring a specific format (which
might seem a good idea anyway) and then forcing it or some other way but
this will work for now. The other variables/fields would not trigger
this problem because two use $(date) and the other is required to be
'open' or 'closed' and nothing else.