Important bug fix in ioccc_status.sh script
The -i version variable was actually the same name as the script version
which means that if one were to use -V to determine the version they
would see not the version of the script but an empty string (unless of
course they first set the version with -i version).

Although this is mostly a cosmetic fix it is still important as the
variables served two very different purposes and should not be used for
both.

Instead the name of the version (variable) of the script was changed to
IOCCC_STATUS_SCRIPT_VERSION. This seemed like a better way to go about
it as the -i version modifies the variable in the status.json file
IOCCC_status_version whereas ioccc_status.sh is the script that does it.
Clever I know :-)